rename kvjournal.rs to kv_journal.rs
