+++ bd-resilient-kv/src/kv_journal.rs
