+++ bd-resilient-kv/src/tests/boundary_test.rs
@@ -7,7 +7,7 @@
 
 // Test cases for performance boundaries and edge cases
 
-use crate::kvjournal::KVJournal;
+use crate::kv_journal::KVJournal;
 use crate::{DoubleBufferedKVJournal, InMemoryKVJournal, MemMappedKVJournal};
 use bd_bonjson::Value;
 use tempfile::TempDir;
