+++ bd-resilient-kv/src/tests/double_buffered_test.rs
@@ -5,7 +5,7 @@
 // LICENSE file or at:
 // https://polyformproject.org/wp-content/uploads/2020/06/PolyForm-Shield-1.0.0.txt
 
-use crate::kvjournal::KVJournal;
+use crate::kv_journal::KVJournal;
 use crate::{DoubleBufferedKVJournal, InMemoryKVJournal};
 use bd_bonjson::Value;
