+++ fuzz/fuzz_targets/resilient_kv_double_buffered.rs
@@ -9,7 +9,7 @@
 
 use arbitrary::{Arbitrary, Unstructured};
 use bd_bonjson::Value;
-use bd_resilient_kv::kvjournal::KVJournal;
+use bd_resilient_kv::kv_journal::KVJournal;
 use bd_resilient_kv::{DoubleBufferedKVJournal, MemMappedKVJournal};
 use std::collections::HashMap;
 use tempfile::TempDir;
