+++ fuzz/fuzz_targets/resilient_kv_in_memory.rs
@@ -10,7 +10,7 @@
 use arbitrary::{Arbitrary, Unstructured};
 use bd_bonjson::Value;
 use bd_resilient_kv::InMemoryKVJournal;
-use bd_resilient_kv::kvjournal::KVJournal;
+use bd_resilient_kv::kv_journal::KVJournal;
 use std::collections::HashMap;
 
 // Wrapper for Value to implement Arbitrary
